requirements.txt
fastapi
uvicorn
python-multipart
openai
torch
transformers
peft
trl
datasets
sentence-transformers
numpy
faiss-cpu
pymupdf
psutil
